Fixed at last
I think.
Fixed external commands hanging.
Fixed a few crashes related to trying to access a dir when its actually a file.
Fixed files being made instead of dirs.
Fixed a crash when running setup a second time while LTS is already set up. It was trying to make a dir when one already exists.
Added some clarification on some options.
Made it so if LTS cleanup crashes, it doesnt delete the cleanup file. Mostly useful when maintaining LTS.
Added the ability to only copy native script files. This makes it easier to maintain LTS.
Removed some redundant text.
Disabled the update checker. It causes a crash because there is an issue with configparser. Will be converted to JSON in the future.
Fixed the python folder not being deleted properly on windows.
Made it so python doesnt get downloaded a second time when LTS is already set up on windows.
Removed some legacy code (checkupdates in commands).
Removed some unused imports.