Fixed Reobf Issue
Fixed an issue with SS that made it so any new classses in net.minecraft.src didn't get moved to the root project on being reobfuscated.

Also removed the ability to select anything other than b1.7.3 in the setup because none of them work.
If anyone wants to make .patch files for them, feel free to do so!