Added b1.7.3-cb-1060
Added b1.7.3-cb-1060. Not fully functional, but with a bit of modification, it would work.

Made it so recompile doesn't push one massive command to the command line. A text file is now generated which is fed to the command line now.

Fixed a crash caused by _winreg not being found. Turns out it is winreg in python 3.

Added a way to easily disable versions by adding a DISABLED file into the version folder.

Setup is now more clear-cut on what you can and can't do.