Added custom python support and some fixes.
Added the ability to use a custom python command to use. It is stored in /runtime/command.

Changed pypy3.6-v7.1.1-linux64 to pypy3.6-7.1.1-beta-linux_x86_64-portable to fix missing dependency issues on linux systems.

Fixed a crash when starting minecraft or a utility crashing on some systems because of an encoding issue.

Fixed EOL on bat and sh files through .gitattributes. They are now CRLF and LF, respectively.

Unified output format.
Made it so setup behaves almost identically between the sh and bat files.

Python version is now logged instead of being printed to console.