Python3 (#21)
Convert to Python3, Java 1.8 compliance, along with initial support for multiple versions.

* Initial, DO NOT MERGE

* Added actual functionality

Still, do not merge.

* Added to setup

Works, but doesnt install any minecraft jars.

* Finished Minecraft install script.

Documentation on how to add stuff to come later.
Script format compliant with Python 2.7=>
Will not run on 3.0=> yet.

* Fixed setup.

installminecraft.py was called twice.
Completely harmless, but still a bug.

* Couple of bugfixes and dropped some "useless" versions.

* Added a1.2.6

Doesnt decompile properly yet.

* Final (?) commit.

* Final commit.

Time to make a PR.

* Python 3

Finally.

* Fixed cleanup

LTS is now returned to an untouched state on cleanup.

* Fixed Linux scripts.

* Final commit

Made the setup look more pretty.
Fixed windows cleanup not being entirely silent.

* Potential setup fix

* Fix extracting PyPy

* Fixed

* Fixed at last

I think.
Fixed external commands hanging.
Fixed a few crashes related to trying to access a dir when its actually a file.
Fixed files being made instead of dirs.
Fixed a crash when running setup a second time while LTS is already set up. It was trying to make a dir when one already exists.
Added some clarification on some options.
Made it so if LTS cleanup crashes, it doesnt delete the cleanup file. Mostly useful when maintaining LTS.
Added the ability to only copy native script files. This makes it easier to maintain LTS.
Removed some redundant text.
Disabled the update checker. It causes a crash because there is an issue with configparser. Will be converted to JSON in the future.
Fixed the python folder not being deleted properly on windows.
Made it so python doesnt get downloaded a second time when LTS is already set up on windows.
Removed some legacy code (checkupdates in commands).
Removed some unused imports.

* Another patch

I'm not even bothering to say fixed anymore.

Merged Dreku's tileentity fix and made it so cleanup doesnt delete python anymore.
Fixed a crash with the logger because the new version of subprocess requires a bytes object to compare.

* Patch

Fixed runclient and runserver crashing.
Changed it so log lines are string and not bytes.
Fixed cleanup not being removed after cleanup.

* Merge New Commit

https://github.com/ModificationStation/1.7.3-LTS/commit/3e67852c9bacd08b3b37df120cfe5a9b66fbdf0c

Merged commit to LTS:python3

* Merge Pull 24 from Master

https://github.com/ModificationStation/1.7.3-LTS/commit/3e67852c9bacd08b3b37df120cfe5a9b66fbdf0c

Merged commit to LTS:python3

* Fixed Reobf Issue

Fixed an issue with SS that made it so any new classses in net.minecraft.src didn't get moved to the root project on being reobfuscated.

Also removed the ability to select anything other than b1.7.3 in the setup because none of them work.
If anyone wants to make .patch files for them, feel free to do so!

* Added Missing Newlines

* Make Linux scripts executable

* Initial macOS compatibility

* Added b1.7.3-cb-1060

Added b1.7.3-cb-1060. Not fully functional, but with a bit of modification, it would work.

Made it so recompile doesn't push one massive command to the command line. A text file is now generated which is fed to the command line now.

Fixed a crash caused by _winreg not being found. Turns out it is winreg in python 3.

Added a way to easily disable versions by adding a DISABLED file into the version folder.

Setup is now more clear-cut on what you can and can't do.

* Fixed jinput.zip name.

* Added custom python support and some fixes.

Added the ability to use a custom python command to use. It is stored in /runtime/command.

Changed pypy3.6-v7.1.1-linux64 to pypy3.6-7.1.1-beta-linux_x86_64-portable to fix missing dependency issues on linux systems.

Fixed a crash when starting minecraft or a utility crashing on some systems because of an encoding issue.

Fixed EOL on bat and sh files through .gitattributes. They are now CRLF and LF, respectively.

Unified output format.
Made it so setup behaves almost identically between the sh and bat files.

Python version is now logged instead of being printed to console.

* Encoding error stopgap

This will definitely fix any encoding-related crashes when logging an external program output.

When an error happens an external program will keep running, but will stop being logged until the external program exits.

* Fixed incorrect setup args

I forgot arrays start at 0 and that the python script itself is passed as an argument.

* Fix startclient

* Fix cleanup being able to be run after decomp

Also slightly clearer logging.

* Eclipse workspace now uses Java 1.8 compliance

Used to use 1.6, but most people use 1.8 now.
Plus: lambdas.